Remove version from InternalUsersModel
