+++ sandbox/modules/identity/src/test/java/org/opensearch/identity/InternalRealmTests.java
@@ -13,53 +13,29 @@ import org.apache.shiro.authc.AuthenticationInfo;
 import org.apache.shiro.authc.IncorrectCredentialsException;
 import org.apache.shiro.authc.UsernamePasswordToken;
 import org.junit.Before;
+import org.opensearch.identity.configuration.SecurityDynamicConfiguration;
 import org.opensearch.identity.configuration.model.InternalUsersModel;
 import org.opensearch.identity.realm.InternalRealm;
 import org.opensearch.identity.realm.InternalUsersStore;
 import org.opensearch.test.OpenSearchTestCase;
 
-import java.util.Map;
-
 public class InternalRealmTests extends OpenSearchTestCase {
 
     private InternalRealm realm;
 
-    private class TestInternalUsersModel extends InternalUsersModel {
-
-        private User user1 = new User("admin", "$2a$12$VcCDgh2NDk07JGN0rjGbM.Ad41qVR/YFJcgHp0UGns5JDymv..TOG", null, null);
-        private final Map<String, User> userModel = Map.of("admin", user1);
-
-        @Override
-        public User getUser(String username) {
-            return userModel.get(username);
-        }
-
-        @Override
-        public boolean exists(String username) {
-            return userModel.containsKey(username);
-        }
-
-        @Override
-        public Map<String, String> getAttributes(String username) {
-            User tmp = userModel.get(username);
-            return tmp == null ? null : tmp.getAttributes();
-        }
-
-        @Override
-        public String getHash(String username) {
-            User tmp = userModel.get(username);
-            return tmp == null ? null : tmp.getHash();
-        }
-    }
-
-    private TestInternalUsersModel ium = new TestInternalUsersModel();
+    private InternalUsersModel ium;
 
     @Before
     public void setUpAndInitializeRealm() throws Exception {
+        SecurityDynamicConfiguration<User> internalUsers = SecurityDynamicConfiguration.empty();
+        User user1 = new User("admin", "$2a$12$VcCDgh2NDk07JGN0rjGbM.Ad41qVR/YFJcgHp0UGns5JDymv..TOG", null, null);
+        internalUsers.putCObject("admin", user1);
+        InternalUsersModel testModel = new InternalUsersModel(internalUsers);
+
         InternalUsersStore.getInstance(); // Instantiates InternalUsersStore
         Thread.sleep(100);
         realm = new InternalRealm.Builder("test").build();
-        InternalUsersStore.getInstance().onInternalUsersModelChanged(ium);
+        InternalUsersStore.getInstance().onInternalUsersModelChanged(testModel);
     }
 
     public void testGetAuthenticationInfoUserExists() {
